Refactor src/components/OrgSettings/ActionItemCategories/OrgActionItemCategories.module.css
