--- src/components/OrgSettings/ActionItemCategories/OrgActionItemCategories.module.css
@@ -1,138 +1,0 @@
-/* Button Styles */
-.addButton {
-  /* Position and size of the button */
-  width: 7em;
-  position: absolute;
-  right: 1rem;
-  top: 1rem;
-}
-
-/* Modal Styles */
-.createModal {
-  /* Position and size of the modal */
-  margin-top: 20vh;
-  margin-left: 13vw;
-  max-width: 80vw;
-}
-
-.icon {
-  /* Size and color of the icon */
-  transform: scale(1.5);
-  color: var(--bs-danger);
-  margin-bottom: 1rem;
-}
-
-.message {
-  /* Centering the content of the modal */
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-.titlemodal {
-  /* Styling for the modal title */
-  color: #707070;
-  font-weight: 600;
-  font-size: 32px;
-  width: 65%;
-  margin-bottom: 0px;
-}
-
-.modalCloseBtn {
-  /* Styling for the modal close button */
-  width: 40px;
-  height: 40px;
-  padding: 1rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-/* Input Styles */
-.noOutline input {
-  /* Removing the outline from the input */
-  outline: none;
-}
-
-/* Header and Action Item Categories Styles */
-.btnsContainer {
-  /* Styling for the container of the buttons */
-  display: flex;
-  margin: 0.5rem 0 1.5rem 0;
-}
-
-.btnsContainer .input {
-  /* Styling for the input field */
-  flex: 1;
-  min-width: 18rem;
-  position: relative;
-}
-
-.btnsContainer input {
-  /* Styling for the input border */
-  outline: 1px solid var(--bs-gray-400);
-}
-
-.btnsContainer .input button {
-  /* Styling for the button in the input field */
-  width: 52px;
-}
-
-.inputField {
-  margin-top: 10px;
-  margin-bottom: 10px;
-  background-color: white;
-  box-shadow: 0 1px 1px #31bb6b;
-}
-
-.inputField > button {
-  padding-top: 10px;
-  padding-bottom: 10px;
-}
-
-/* Dropdown Styles */
-.dropdown {
-  /* Styling for the dropdown */
-  background-color: white;
-  border: 1px solid #31bb6b;
-  position: relative;
-  display: inline-block;
-  color: #31bb6b;
-}
-
-/* Datagrid Styles */
-.rowBackground {
-  /* Styling for the row background */
-  background-color: var(--bs-white);
-  max-height: 120px;
-}
-
-.tableHeader {
-  /* Styling for the table header */
-  background-color: var(--bs-primary);
-  color: var(--bs-white);
-  font-size: 1rem;
-}
-
-.chipIcon {
-  /* Styling for the chip icon */
-  height: 0.9rem !important;
-}
-
-.chip {
-  /* Styling for the chip */
-  height: 1.5rem !important;
-}
-
-.active {
-  /* Styling for the active state */
-  background-color: #31bb6a50 !important;
-}
-
-.pending {
-  /* Styling for the pending state */
-  background-color: #ffd76950 !important;
-  color: #bb952bd0 !important;
-  border-color: #bb952bd0 !important;
-}
